pyproject.toml
[project]
name = "inspection-ai-pilot"
version = "0.1.0"
description = "工程检测智能领航员"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "faiss-cpu>=1.7.4",
    "langchain>=0.3.0,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "langchain-huggingface>=0.1.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.32.0",
    "watchdog>=4.0.0",
    "dashscope>=1.14.0",
    "sniffio>=1.3.0",
    "httpx>=0.27.0",
    "huggingface_hub>=0.23.0",
    "pytest>=8.0.0",
    "sqlalchemy>=2.0.45",
    "pandas>=2.3.3",
]

[tool.uv]
# 可以在这里锁定具体的 Python 版本，防止在 3.13 上运行出错
package = true
